tests/guardrail_litellm/requirements.txt
litellm
fastmcp
pytest
python-dotenv
google-generativeai
uvicorn
fastapi
